adding long long overloads to basic ops, and fixing general size_t issues on windows x64
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1012 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/io/image.h
@@ -224,8 +224,8 @@ private:
 		float exposure;
 		float colorCorr[sizeColorCorr];
 		float primaries[sizePrimaries];
-		unsigned height;
-		unsigned width;
+		size_t height;
+		size_t width;
 		bool yIncreasing;
 		bool xIncreasing;
 		bool isRgb;
